rename payload types file
