+++ src/electron/flux/action/device-actions.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { Action } from 'common/flux/action';
-import { UpdateDevicePayload } from './update-device-payload';
+import { UpdateDevicePayload } from './device-action-payloads';
 
 export class DeviceActions {
     public readonly updateDevice = new Action<UpdateDevicePayload>();
